Pass users to reviews
