--- api_yamdb/users/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from .models import User
-
-admin.site.register(User)
