--- api_yamdb/users/models.py
@@ -1,23 +1,0 @@
-from django.contrib.auth.models import AbstractUser
-from django.db import models
-
-
-class User(AbstractUser):
-    USER = 'us'
-    MODERATOR = 'mo'
-    ADMIN = 'ad'
-    ROLE_CHOICES = (
-        (USER, 'user',),
-        (MODERATOR, 'moderator'),
-        (ADMIN, 'admin')
-    )
-    email = models.EmailField(unique=True)
-    role = models.CharField(
-        max_length=2,
-        choices=ROLE_CHOICES,
-        default=USER,
-    )
-    bio = models.TextField(
-        'Биография',
-        blank=True,
-    )
